usb: gadget: mbim: Remove QBI block on usb cable disconnection
QBI is a user space module that transfers control packets between
USB and modem.

On USB cable disconnection QBI was blocked, and didn't receive a
notification on the cable disconnection, so QBI couldn't notify the modem
on this event as well. Once we connect the USB cable the QBI receive both
disconnect message and connect message and notify the modem. This can
result in a situation where QBI goes out of sync with the current state
of the system.

By this fix, we don't block the QBI on cable disconnection, so USB will
notify QBI on the disconnection allowing QBI to stay in sync with the
system state.

Patch edited for Sierra mdm9x15 integration.

CRs-Fixed: 521610
Change-Id: I6779e94570bf8848307cc77355c96e485cd65e52
Signed-off-by: Bar Weiner <bweiner@codeaurora.org>
Signed-off-by: Mihai Dragusu <mihai.dragusu@windriver.com>