msm_otg: suspend/resume on cable unplug/plug
The expected behaviour of Legato is to suspend when the Head Unit
connected over the USB OTG interface disconnects or performs a
runtime suspend of its USB peripheral. This patch implements the
desired functionality in conformance to Android autosleep.

The msm_otg device has a wakeup source that is taken whenever the
USB OTG detects a standard upstream charger, and releases it when
the upstream host decides to suspend its peripheral, which allows
autosleep to kick-in and suspend the device.

This commit blatantly assumes a certain type of upstream device.
In reality there may be multiple options of powering the USB OTG
interface, as well as negotiation between endpoints on who takes
the role of host and peripheral. The ultimate decision on when to
suspend and resume needs to be brought as result of negotiation,
and should also involve user-space application.

FIXME: This patch addresses the immediate requirement and should
be removed when a comprehensive solution to USB OTG interface is
implemented.

Change-Id: I864b9d79fb43e33a9796e9bb06c641d91b603b11
Signed-off-by: Zoran Markovic <zmarkovic@sierrawireless.com>
Signed-off-by: Dragan Marinkovic <dmarinkovi@sierrawireless.com>