msm_otg: Added epoll interface for EPOLLWAKEUP
Created misc device interface so that the USB OTG interface can be
polled for change of link status. Applications that require AP to
be awake when link is active should perform epoll_wait() with
EPOLLWAKEUP flag enabled on /dev/usb_link device to ensure AP is
woken up when link status is changed.  To retrieve the status of
link, perform a read() of /dev/usb_link device as follows:
        int link_status;
        read(fd, &link_status, sizeof(link_status));

Valid link_status values are 0 (suspended) and 1 (active).

For now only one file descriptor can be open on /dev/usb_link.
Implementation is likely to undergo more refinements.

Change-Id: I99a4d624ec891f31bbb3c0631c42ebf12d729e02
Signed-off-by: Zoran Markovic <zmarkovic@sierrawireless.com>